+++ doc_examples/guide/errors/error_handlers/project/src/blueprint.rs
@@ -3,5 +3,6 @@ use pavex::blueprint::Blueprint;
 pub fn blueprint() -> Blueprint {
     let mut bp = Blueprint::new();
     bp.nest(crate::core::blueprint());
+    bp.nest(crate::universal::blueprint());
     bp
 }
