+++ doc_examples/guide/errors/error_handlers/project/src/core/error_handler.rs
@@ -1,6 +1,6 @@
 use pavex::http::StatusCode;
 
-pub async fn login_error2response(e: &LoginError) -> StatusCode /* (1)! */ {
+pub async fn login_error2response(e: &LoginError) -> StatusCode {
     match e {
         LoginError::InvalidCredentials => StatusCode::UNAUTHORIZED,
         LoginError::DatabaseError => StatusCode::INTERNAL_SERVER_ERROR,
