+++ doc_examples/guide/errors/error_handlers/project/src/lib.rs
@@ -4,3 +4,4 @@ pub use blueprint::blueprint;
 
 mod blueprint;
 pub mod core;
+pub mod universal;
