+++ doc_examples/guide/errors/error_handlers/project/src/universal/blueprint.rs
@@ -1,0 +1,10 @@
+use pavex::blueprint::router::POST;
+use pavex::blueprint::Blueprint;
+use pavex::f;
+
+pub fn blueprint() -> Blueprint {
+    let mut bp = Blueprint::new();
+    bp.route(POST, "/login", f!(crate::core::handler))
+        .error_handler(f!(crate::core::login_error2response)); // (1)!
+    bp
+}
