+++ doc_examples/guide/errors/error_handlers/project/src/universal/mod.rs
@@ -1,0 +1,7 @@
+pub use blueprint::blueprint;
+pub use error_handler::login_error2response;
+pub use routes::handler;
+
+mod blueprint;
+mod error_handler;
+mod routes;
