+++ doc_examples/guide/errors/error_handlers/project/src/universal/routes.rs
