+++ libs/pavex_cli/tests/ui_tests/app_builder/expectations/diagnostics.dot
@@ -8,19 +8,19 @@ digraph "GET /home - 0" {
     6 [ label = "<pavex::response::Response as pavex::response::IntoResponse>::into_response(pavex::response::Response) -> pavex::response::Response"]
     7 [ label = "core::prelude::rust_2015::Result<std::path::PathBuf, app::ExtractPathError<alloc::string::String>> -> app::ExtractPathError<alloc::string::String>"]
     8 [ label = "app::handle_extract_path_error(&app::ExtractPathError<alloc::string::String>, app::Logger) -> pavex::response::Response"]
-    9 [ label = "<pavex::response::Response as pavex::response::IntoResponse>::into_response(pavex::response::Response) -> pavex::response::Response"]
-    10 [ label = "app::logger() -> app::Logger"]
+    9 [ label = "app::logger() -> app::Logger"]
+    10 [ label = "<pavex::response::Response as pavex::response::IntoResponse>::into_response(pavex::response::Response) -> pavex::response::Response"]
     11 [ label = "`match`"]
     3 -> 0 [ ]
     2 -> 0 [ ]
     1 -> 0 [ ]
     11 -> 7 [ ]
     5 -> 4 [ ]
-    0 -> 6 [ ]
     11 -> 1 [ ]
-    8 -> 9 [ ]
+    0 -> 6 [ ]
     7 -> 8 [ label = "&"]
-    10 -> 8 [ ]
+    9 -> 8 [ ]
+    8 -> 10 [ ]
     4 -> 11 [ ]
 }
