+++ libs/pavex_cli/tests/ui_tests/blueprint/common/components_can_fail/expectations/diagnostics.dot
@@ -6,18 +6,20 @@ digraph "GET /home - 0" {
     4 [ label = "pavex::request::RequestHead"]
     5 [ label = "core::prelude::rust_2015::Result<pavex::response::Response, app::MiddlewareError> -> pavex::response::Response"]
     6 [ label = "core::prelude::rust_2015::Result<pavex::response::Response, app::MiddlewareError> -> app::MiddlewareError"]
-    7 [ label = "app::handle_middleware_error(&app::MiddlewareError) -> pavex::response::Response"]
-    8 [ label = "<pavex::response::Response as pavex::response::IntoResponse>::into_response(pavex::response::Response) -> pavex::response::Response"]
-    9 [ label = "`match`"]
+    7 [ label = "<pavex::response::Response as pavex::response::IntoResponse>::into_response(pavex::response::Response) -> pavex::response::Response"]
+    8 [ label = "app::handle_middleware_error(&app::MiddlewareError) -> pavex::response::Response"]
+    9 [ label = "<pavex::response::Response as pavex::response::IntoResponse>::into_response(pavex::response::Response) -> pavex::response::Response"]
+    10 [ label = "`match`"]
     1 -> 0 [ ]
     2 -> 1 [ ]
     4 -> 2 [ ]
     3 -> 2 [ ]
-    9 -> 6 [ ]
-    9 -> 5 [ ]
-    7 -> 8 [ ]
-    6 -> 7 [ label = "&"]
-    0 -> 9 [ ]
+    10 -> 5 [ ]
+    10 -> 6 [ ]
+    6 -> 8 [ label = "&"]
+    5 -> 7 [ ]
+    8 -> 9 [ ]
+    0 -> 10 [ ]
 }
 
 digraph "GET /home - 1" {
@@ -32,15 +34,15 @@ digraph "GET /home - 1" {
     8 [ label = "core::prelude::rust_2015::Result<pavex::response::Response, app::HandlerError> -> app::HandlerError"]
     9 [ label = "core::prelude::rust_2015::Result<app::Logger, app::LoggerError> -> app::LoggerError"]
     10 [ label = "core::prelude::rust_2015::Result<std::path::PathBuf, app::ExtractPathError<alloc::string::String>> -> app::ExtractPathError<alloc::string::String>"]
-    11 [ label = "app::handle_handler_error(&app::HandlerError) -> pavex::response::Response"]
-    12 [ label = "app::handle_logger_error(&app::LoggerError) -> pavex::response::Response"]
-    13 [ label = "app::handle_extract_path_error(&app::ExtractPathError<alloc::string::String>, app::Logger) -> pavex::response::Response"]
-    14 [ label = "<pavex::response::Response as pavex::response::IntoResponse>::into_response(pavex::response::Response) -> pavex::response::Response"]
-    15 [ label = "<pavex::response::Response as pavex::response::IntoResponse>::into_response(pavex::response::Response) -> pavex::response::Response"]
-    16 [ label = "<pavex::response::Response as pavex::response::IntoResponse>::into_response(pavex::response::Response) -> pavex::response::Response"]
+    11 [ label = "<pavex::response::Response as pavex::response::IntoResponse>::into_response(pavex::response::Response) -> pavex::response::Response"]
+    12 [ label = "app::handle_handler_error(&app::HandlerError) -> pavex::response::Response"]
+    13 [ label = "app::handle_logger_error(&app::LoggerError) -> pavex::response::Response"]
+    14 [ label = "app::handle_extract_path_error(&app::ExtractPathError<alloc::string::String>, app::Logger) -> pavex::response::Response"]
+    15 [ label = "core::prelude::rust_2015::Result<app::Logger, app::LoggerError> -> app::Logger"]
+    16 [ label = "app::logger() -> core::prelude::rust_2015::Result<app::Logger, app::LoggerError>"]
     17 [ label = "<pavex::response::Response as pavex::response::IntoResponse>::into_response(pavex::response::Response) -> pavex::response::Response"]
-    18 [ label = "core::prelude::rust_2015::Result<app::Logger, app::LoggerError> -> app::Logger"]
-    19 [ label = "app::logger() -> core::prelude::rust_2015::Result<app::Logger, app::LoggerError>"]
+    18 [ label = "<pavex::response::Response as pavex::response::IntoResponse>::into_response(pavex::response::Response) -> pavex::response::Response"]
+    19 [ label = "<pavex::response::Response as pavex::response::IntoResponse>::into_response(pavex::response::Response) -> pavex::response::Response"]
     20 [ label = "core::prelude::rust_2015::Result<app::Logger, app::LoggerError> -> app::LoggerError"]
     21 [ label = "app::handle_logger_error(&app::LoggerError) -> pavex::response::Response"]
     22 [ label = "<pavex::response::Response as pavex::response::IntoResponse>::into_response(pavex::response::Response) -> pavex::response::Response"]
@@ -54,26 +56,26 @@ digraph "GET /home - 1" {
     1 -> 0 [ ]
     25 -> 10 [ ]
     6 -> 5 [ ]
-    23 -> 8 [ ]
-    23 -> 7 [ ]
-    24 -> 2 [ ]
     25 -> 1 [ ]
-    7 -> 14 [ ]
-    11 -> 15 [ ]
-    12 -> 16 [ ]
-    13 -> 17 [ ]
-    10 -> 13 [ label = "&"]
-    18 -> 13 [ ]
+    24 -> 2 [ ]
+    23 -> 7 [ ]
+    23 -> 8 [ ]
+    10 -> 14 [ label = "&"]
+    15 -> 14 [ ]
     26 -> 20 [ ]
-    9 -> 12 [ label = "&"]
-    8 -> 11 [ label = "&"]
-    26 -> 18 [ ]
-    21 -> 22 [ ]
+    9 -> 13 [ label = "&"]
+    8 -> 12 [ label = "&"]
+    7 -> 11 [ ]
+    26 -> 15 [ ]
+    14 -> 19 [ ]
+    13 -> 18 [ ]
+    12 -> 17 [ ]
     20 -> 21 [ label = "&"]
+    21 -> 22 [ ]
     0 -> 23 [ ]
     4 -> 24 [ ]
     5 -> 25 [ ]
-    19 -> 26 [ ]
+    16 -> 26 [ ]
 }
 
 digraph "* /home - 0" {
@@ -82,18 +84,20 @@ digraph "* /home - 0" {
     2 [ label = "crate::route_1::Next0(&'a pavex::router::AllowedMethods) -> crate::route_1::Next0<'a>"]
     4 [ label = "core::prelude::rust_2015::Result<pavex::response::Response, app::MiddlewareError> -> pavex::response::Response"]
     5 [ label = "core::prelude::rust_2015::Result<pavex::response::Response, app::MiddlewareError> -> app::MiddlewareError"]
-    6 [ label = "app::handle_middleware_error(&app::MiddlewareError) -> pavex::response::Response"]
-    7 [ label = "<pavex::response::Response as pavex::response::IntoResponse>::into_response(pavex::response::Response) -> pavex::response::Response"]
-    8 [ label = "`match`"]
-    9 [ label = "&pavex::router::AllowedMethods"]
+    6 [ label = "<pavex::response::Response as pavex::response::IntoResponse>::into_response(pavex::response::Response) -> pavex::response::Response"]
+    7 [ label = "app::handle_middleware_error(&app::MiddlewareError) -> pavex::response::Response"]
+    8 [ label = "<pavex::response::Response as pavex::response::IntoResponse>::into_response(pavex::response::Response) -> pavex::response::Response"]
+    9 [ label = "`match`"]
+    10 [ label = "&pavex::router::AllowedMethods"]
     1 -> 0 [ ]
     2 -> 1 [ ]
-    8 -> 5 [ ]
-    8 -> 4 [ ]
-    6 -> 7 [ ]
-    5 -> 6 [ label = "&"]
-    0 -> 8 [ ]
-    9 -> 2 [ ]
+    9 -> 4 [ ]
+    9 -> 5 [ ]
+    5 -> 7 [ label = "&"]
+    4 -> 6 [ ]
+    7 -> 8 [ ]
+    0 -> 9 [ ]
+    10 -> 2 [ ]
 }
 
 digraph "* /home - 1" {
@@ -117,8 +121,8 @@ digraph app_state {
     1 -> 0 [ ]
     8 -> 5 [ ]
     3 -> 2 [ ]
-    0 -> 4 [ ]
     8 -> 1 [ ]
+    0 -> 4 [ ]
     5 -> 6 [ ]
     6 -> 7 [ ]
     2 -> 8 [ ]
