+++ libs/pavex_cli/tests/ui_tests/blueprint/constructors/output_type_of_constructors_can_contain_generic_parameters/expectations/diagnostics.dot
@@ -39,20 +39,20 @@ digraph "GET /home - 0" {
     3 -> 0 [ label = "&"]
     2 -> 0 [ ]
     1 -> 0 [ ]
-    0 -> 12 [ ]
-    25 -> 7 [ ]
-    26 -> 6 [ ]
-    27 -> 5 [ ]
     28 -> 4 [ ]
-    17 -> 21 [ ]
-    18 -> 22 [ ]
-    19 -> 23 [ ]
-    20 -> 24 [ ]
+    27 -> 5 [ ]
+    26 -> 6 [ ]
+    25 -> 7 [ ]
+    0 -> 12 [ ]
     16 -> 20 [ label = "&"]
     15 -> 19 [ label = "&"]
     14 -> 18 [ label = "&"]
     13 -> 17 [ label = "&"]
     1 -> 17 [ label = "&"]
+    20 -> 24 [ ]
+    19 -> 23 [ ]
+    18 -> 22 [ ]
+    17 -> 21 [ ]
     8 -> 25 [ ]
     9 -> 26 [ ]
     10 -> 27 [ ]
