+++ libs/pavex_cli/tests/ui_tests/blueprint/error_handlers/error_handlers_can_take_pavex_error_rather_than_specific_error_type/test_config.toml
@@ -1,0 +1,6 @@
+description = """Pavex allows an error handler to take &pavex::Error as input
+rather than the specific error type returned by the fallible component
+it is associated with."""
+
+[expectations]
+codegen = "pass"
