+++ libs/pavex_cli/tests/ui_tests/borrow_checker/control_flow/you_can_consume_a_non_cloneable_type_from_two_different_control_flow_branches/expectations/diagnostics.dot
@@ -11,11 +11,11 @@ digraph "GET /home - 0" {
     2 -> 0 [ ]
     1 -> 0 [ ]
     8 -> 5 [ ]
-    0 -> 4 [ ]
     8 -> 1 [ ]
-    6 -> 7 [ ]
+    0 -> 4 [ ]
     5 -> 6 [ label = "&"]
     2 -> 6 [ ]
+    6 -> 7 [ ]
     3 -> 8 [ ]
 }
