+++ libs/pavex_cli/tests/ui_tests/dependency_injection/clone_nodes_are_only_added_where_needed/expectations/diagnostics.dot
@@ -3,12 +3,14 @@ digraph "GET / - 0" {
     1 [ label = "app::Singleton"]
     2 [ label = "pavex::middleware::Next::new(crate::route_0::Next0) -> pavex::middleware::Next<crate::route_0::Next0>"]
     3 [ label = "crate::route_0::Next0(app::Singleton) -> crate::route_0::Next0"]
-    4 [ label = "<app::Singleton as core::clone::Clone>::clone(&app::Singleton) -> app::Singleton"]
+    4 [ label = "<pavex::response::Response as pavex::response::IntoResponse>::into_response(pavex::response::Response) -> pavex::response::Response"]
+    5 [ label = "<app::Singleton as core::clone::Clone>::clone(&app::Singleton) -> app::Singleton"]
     2 -> 0 [ ]
     3 -> 2 [ ]
     1 -> 3 [ ]
-    1 -> 4 [ label = "&"]
-    4 -> 0 [ ]
+    0 -> 4 [ ]
+    1 -> 5 [ label = "&"]
+    5 -> 0 [ ]
 }
 
 digraph "GET / - 1" {
@@ -24,11 +26,13 @@ digraph "* / - 0" {
     1 [ label = "app::Singleton"]
     2 [ label = "pavex::middleware::Next::new(crate::route_1::Next0<'a>) -> pavex::middleware::Next<crate::route_1::Next0<'a>>"]
     3 [ label = "crate::route_1::Next0(&'a pavex::router::AllowedMethods) -> crate::route_1::Next0<'a>"]
-    5 [ label = "&pavex::router::AllowedMethods"]
+    5 [ label = "<pavex::response::Response as pavex::response::IntoResponse>::into_response(pavex::response::Response) -> pavex::response::Response"]
+    6 [ label = "&pavex::router::AllowedMethods"]
     2 -> 0 [ ]
     3 -> 2 [ ]
     1 -> 0 [ ]
-    5 -> 3 [ ]
+    0 -> 5 [ ]
+    6 -> 3 [ ]
 }
 
 digraph "* / - 1" {
