+++ libs/pavex_cli/tests/ui_tests/error_observers/error_observer_with_generic_error/expectations/diagnostics.dot
@@ -4,19 +4,19 @@ digraph "GET /home - 0" {
     2 [ label = "app::constructor() -> core::prelude::rust_2015::Result<app::Generic<alloc::string::String>, app::GenericError<alloc::string::String>>"]
     3 [ label = "<pavex::response::Response as pavex::response::IntoResponse>::into_response(pavex::response::Response) -> pavex::response::Response"]
     4 [ label = "core::prelude::rust_2015::Result<app::Generic<alloc::string::String>, app::GenericError<alloc::string::String>> -> app::GenericError<alloc::string::String>"]
-    5 [ label = "app::error_handler(&app::GenericError<alloc::string::String>) -> pavex::response::Response"]
-    6 [ label = "pavex::Error::new(app::GenericError<alloc::string::String>) -> pavex::Error"]
+    5 [ label = "pavex::Error::new(app::GenericError<alloc::string::String>) -> pavex::Error"]
+    6 [ label = "app::error_handler(&app::GenericError<alloc::string::String>) -> pavex::response::Response"]
     7 [ label = "<pavex::response::Response as pavex::response::IntoResponse>::into_response(pavex::response::Response) -> pavex::response::Response"]
     8 [ label = "app::error_observer(&pavex::Error)"]
     9 [ label = "`match`"]
     1 -> 0 [ ]
     9 -> 4 [ ]
-    0 -> 3 [ ]
     9 -> 1 [ ]
-    4 -> 6 [ ]
-    5 -> 7 [ ]
-    4 -> 5 [ label = "&"]
-    6 -> 8 [ label = "&"]
+    0 -> 3 [ ]
+    4 -> 6 [ label = "&"]
+    4 -> 5 [ ]
+    6 -> 7 [ ]
+    5 -> 8 [ label = "&"]
     8 -> 7 [ label = "happens before"]
     2 -> 9 [ ]
 }
