+++ libs/pavex_cli/tests/ui_tests/error_observers/error_observers_can_depend_on_fallible_singletons/expectations/diagnostics.dot
@@ -4,24 +4,24 @@ digraph "GET /home - 0" {
     2 [ label = "app::b(&app::A) -> core::prelude::rust_2015::Result<app::B, app::AnError>"]
     4 [ label = "<pavex::response::Response as pavex::response::IntoResponse>::into_response(pavex::response::Response) -> pavex::response::Response"]
     5 [ label = "core::prelude::rust_2015::Result<app::B, app::AnError> -> app::AnError"]
-    6 [ label = "app::error_handler(&app::A, &app::AnError) -> pavex::response::Response"]
-    7 [ label = "pavex::Error::new(app::AnError) -> pavex::Error"]
+    6 [ label = "pavex::Error::new(app::AnError) -> pavex::Error"]
+    7 [ label = "app::error_handler(&app::A, &app::AnError) -> pavex::response::Response"]
     8 [ label = "<pavex::response::Response as pavex::response::IntoResponse>::into_response(pavex::response::Response) -> pavex::response::Response"]
     9 [ label = "app::error_observer(&app::A, &pavex::Error)"]
     10 [ label = "`match`"]
     11 [ label = "&app::A"]
     1 -> 0 [ label = "&"]
     10 -> 5 [ ]
-    0 -> 4 [ ]
     10 -> 1 [ ]
-    5 -> 7 [ ]
-    6 -> 8 [ ]
-    5 -> 6 [ label = "&"]
-    7 -> 9 [ label = "&"]
+    0 -> 4 [ ]
+    5 -> 7 [ label = "&"]
+    5 -> 6 [ ]
+    7 -> 8 [ ]
+    6 -> 9 [ label = "&"]
     9 -> 8 [ label = "happens before"]
     2 -> 10 [ ]
     11 -> 9 [ ]
-    11 -> 6 [ ]
+    11 -> 7 [ ]
     11 -> 2 [ ]
 }
 
@@ -44,8 +44,8 @@ digraph app_state {
     7 [ label = "`match`"]
     1 -> 0 [ ]
     7 -> 4 [ ]
-    0 -> 3 [ ]
     7 -> 1 [ ]
+    0 -> 3 [ ]
     4 -> 5 [ ]
     5 -> 6 [ ]
     2 -> 7 [ ]
