+++ libs/pavex_cli/tests/ui_tests/error_observers/error_observers_happy_path/expectations/diagnostics.dot
@@ -5,8 +5,8 @@ digraph "GET /home - 0" {
     3 [ label = "app::a() -> app::A"]
     4 [ label = "<pavex::response::Response as pavex::response::IntoResponse>::into_response(pavex::response::Response) -> pavex::response::Response"]
     5 [ label = "core::prelude::rust_2015::Result<app::B, app::ErrorB> -> app::ErrorB"]
-    6 [ label = "app::error_handler(&app::A, &app::ErrorB) -> pavex::response::Response"]
-    7 [ label = "pavex::Error::new(app::ErrorB) -> pavex::Error"]
+    6 [ label = "pavex::Error::new(app::ErrorB) -> pavex::Error"]
+    7 [ label = "app::error_handler(&app::A, &app::ErrorB) -> pavex::response::Response"]
     8 [ label = "<pavex::response::Response as pavex::response::IntoResponse>::into_response(pavex::response::Response) -> pavex::response::Response"]
     9 [ label = "app::error_observer(app::A, &pavex::Error)"]
     10 [ label = "app::error_observer2(&pavex::Error)"]
@@ -14,15 +14,15 @@ digraph "GET /home - 0" {
     1 -> 0 [ label = "&"]
     11 -> 5 [ ]
     3 -> 2 [ label = "&"]
-    0 -> 4 [ ]
     11 -> 1 [ ]
-    5 -> 7 [ ]
-    6 -> 8 [ ]
-    5 -> 6 [ label = "&"]
-    3 -> 6 [ label = "&"]
-    7 -> 9 [ label = "&"]
+    0 -> 4 [ ]
+    5 -> 7 [ label = "&"]
+    3 -> 7 [ label = "&"]
+    5 -> 6 [ ]
+    7 -> 8 [ ]
+    6 -> 9 [ label = "&"]
     9 -> 10 [ label = "happens before"]
-    7 -> 10 [ label = "&"]
+    6 -> 10 [ label = "&"]
     10 -> 8 [ label = "happens before"]
     3 -> 9 [ ]
     2 -> 11 [ ]
