+++ libs/pavex_cli/tests/ui_tests/path_parameters/path_parameters_happy_path/expectations/diagnostics.dot
@@ -11,10 +11,10 @@ digraph "GET /home/:home_id - 0" {
     1 -> 0 [ ]
     8 -> 5 [ ]
     3 -> 2 [ ]
-    0 -> 4 [ ]
     8 -> 1 [ ]
-    6 -> 7 [ ]
+    0 -> 4 [ ]
     5 -> 6 [ label = "&"]
+    6 -> 7 [ ]
     2 -> 8 [ ]
 }
 
@@ -39,10 +39,10 @@ digraph "GET /home/:home_id/room/:room_id - 0" {
     1 -> 0 [ ]
     8 -> 5 [ ]
     3 -> 2 [ ]
-    0 -> 4 [ ]
     8 -> 1 [ ]
-    6 -> 7 [ ]
+    0 -> 4 [ ]
     5 -> 6 [ label = "&"]
+    6 -> 7 [ ]
     2 -> 8 [ ]
 }
 
@@ -67,10 +67,10 @@ digraph "GET /town/*town - 0" {
     1 -> 0 [ ]
     8 -> 5 [ ]
     3 -> 2 [ ]
-    0 -> 4 [ ]
     8 -> 1 [ ]
-    6 -> 7 [ ]
+    0 -> 4 [ ]
     5 -> 6 [ label = "&"]
+    6 -> 7 [ ]
     2 -> 8 [ ]
 }
