+++ libs/pavexc/src/compiler/analyses/call_graph/borrow_checker/clone.rs
@@ -5,8 +5,9 @@ use once_cell::sync::OnceCell;
 
 use pavex_bp_schema::CloningStrategy;
 
+use crate::compiler::analyses::components::{ComponentDb, ComponentId};
 use crate::compiler::analyses::components::{
-    ComponentDb, ComponentId, ConsumptionMode, HydratedComponent, InsertTransformer,
+    ConsumptionMode, HydratedComponent, InsertTransformer,
 };
 use crate::compiler::analyses::computations::ComputationDb;
 use crate::compiler::analyses::user_components::ScopeId;
@@ -93,6 +94,7 @@ pub(super) fn get_clone_component_id(
         scope_id,
         InsertTransformer::Lazily,
         ConsumptionMode::SharedBorrow,
+        0,
         computation_db,
     );
     Some(clone_component_id)
